refactored build in example to current architecture and adjusted api route
