+++ ShortcutTrainerBackend/ShortcutTrainerBackend/Controllers/WeatherForecastController.cs
@@ -1,32 +1,25 @@
 using Microsoft.AspNetCore.Mvc;
+using ShortcutTrainerBackend.Data.Models;
+using ShortcutTrainerBackend.Services.Interfaces;
 
 namespace ShortcutTrainerBackend.Controllers;
 
 [ApiController]
-[Route("[controller]")]
+[Route("api/[controller]")]
 public class WeatherForecastController : ControllerBase
 {
-    private static readonly string[] Summaries = new[]
-    {
-        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-    };
-
-    private readonly ILogger<WeatherForecastController> _logger;
-
-    public WeatherForecastController(ILogger<WeatherForecastController> logger)
+    public WeatherForecastController(ILogger<WeatherForecastController> logger, IWeatherForecastService weatherForecastService)
     {
         _logger = logger;
+        _weatherForecastService = weatherForecastService;
     }
 
+    private readonly ILogger<WeatherForecastController> _logger;
+    private readonly IWeatherForecastService _weatherForecastService;
+
     [HttpGet(Name = "GetWeatherForecast")]
-    public IEnumerable<WeatherForecast> Get()
+    public async Task<IEnumerable<WeatherForecast>> Get()
     {
-        return Enumerable.Range(1, 5).Select(index => new WeatherForecast
-            {
-                Date = DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
-                TemperatureC = Random.Shared.Next(-20, 55),
-                Summary = Summaries[Random.Shared.Next(Summaries.Length)]
-            })
-            .ToArray();
+        return await _weatherForecastService.GetRandomWeatherForecastAsync();
     }
 }
