+++ ShortcutTrainerBackend/ShortcutTrainerBackend/Data/Models/WeatherForecast.cs
@@ -1,4 +1,4 @@
-namespace ShortcutTrainerBackend;
+namespace ShortcutTrainerBackend.Data.Models;
 
 public class WeatherForecast
 {
