+++ ShortcutTrainerBackend/ShortcutTrainerBackend/Services/Interfaces/IWeatherForecastService.cs
@@ -1,0 +1,8 @@
+using ShortcutTrainerBackend.Data.Models;
+
+namespace ShortcutTrainerBackend.Services.Interfaces;
+
+public interface IWeatherForecastService
+{
+    Task<IEnumerable<WeatherForecast>> GetRandomWeatherForecastAsync();
+}
