+++ ShortcutTrainerBackend/ShortcutTrainerBackend/Services/JokeService.cs
@@ -1,4 +1,3 @@
-using System.Collections.ObjectModel;
 using ShortcutTrainerBackend.Data.Models;
 using ShortcutTrainerBackend.Services.Interfaces;
 using ShortcutTrainerBackend.Testing.Mocks.Interfaces;
