+++ ShortcutTrainerBackend/ShortcutTrainerBackend/Services/WeatherForecastService.cs
@@ -1,0 +1,24 @@
+using ShortcutTrainerBackend.Data.Models;
+using ShortcutTrainerBackend.Services.Interfaces;
+
+namespace ShortcutTrainerBackend.Services;
+
+public class WeatherForecastService : IWeatherForecastService
+{
+    private static readonly string[] Summaries = new[]
+    {
+        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+    };
+
+    public async Task<IEnumerable<WeatherForecast>> GetRandomWeatherForecastAsync()
+    {
+        await Task.Delay(1000); // wait 1 seconds before returning any result
+        return Enumerable.Range(1, 5).Select(index => new WeatherForecast
+            {
+                Date = DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
+                TemperatureC = Random.Shared.Next(-20, 55),
+                Summary = Summaries[Random.Shared.Next(Summaries.Length)]
+            })
+            .ToArray();
+    }
+}
